Sync class list before pushing groups
if a user is added to the class late and the instructor adds them to a
group before pushing the updated class list to iPeer from Blackboard.
If they then try to push groups to iPeer it won’t add the new student
to their group on the iPeer side unless the roster is pushed first.

fix #16